+++ Kernel/System/DynamicField/ObjectType/Ticket.pm
@@ -17,6 +17,7 @@ use Scalar::Util;
 use Kernel::System::VariableCheck qw(:all);
 
 our @ObjectDependencies = (
+    'Kernel::System::DB',
     'Kernel::System::DynamicField::Backend',
     'Kernel::System::Log',
     'Kernel::System::Ticket',
@@ -97,6 +98,16 @@ sub PostValueSet {
             return;
         }
     }
+
+    # get database object
+    my $DBObject = $Kernel::OM->Get('Kernel::System::DB');
+
+    # update change time
+    return if !$DBObject->Do(
+        SQL => 'UPDATE ticket SET change_time = current_timestamp, '
+            . ' change_by = ? WHERE id = ?',
+        Bind => [ \$Param{UserID}, \$Param{ObjectID} ],
+    );
 
     # get ticket object
     my $TicketObject = $Kernel::OM->Get('Kernel::System::Ticket');