+++ Kernel/System/Stats/Dynamic/TicketAccountedTime.pm
@@ -287,6 +287,32 @@ sub GetObjectAttributes {
             },
         },
         {
+            Name             => 'Last changed times',
+            UseAsXvalue      => 1,
+            UseAsValueSeries => 1,
+            UseAsRestriction => 1,
+            Element          => 'LastChangeTime',
+            TimePeriodFormat => 'DateInputFormat',      # 'DateInputFormatLong',
+            Block            => 'Time',
+            Values           => {
+                TimeStart => 'TicketLastChangeTimeNewerDate',
+                TimeStop  => 'TicketLastChangeTimeOlderDate',
+            },
+        },
+        {
+            Name             => 'Change times',
+            UseAsXvalue      => 1,
+            UseAsValueSeries => 1,
+            UseAsRestriction => 1,
+            Element          => 'ChangeTime',
+            TimePeriodFormat => 'DateInputFormat',      # 'DateInputFormatLong',
+            Block            => 'Time',
+            Values           => {
+                TimeStart => 'TicketChangeTimeNewerDate',
+                TimeStop  => 'TicketChangeTimeOlderDate',
+            },
+        },
+        {
             Name             => 'Ticket Close Time',
             UseAsXvalue      => 1,
             UseAsValueSeries => 1,
@@ -1303,6 +1329,8 @@ sub _AllowedTicketSearchAttributes {
         TicketCreateTimeOlderDate
         TicketChangeTimeNewerDate
         TicketChangeTimeOlderDate
+        TicketLastChangeTimeNewerDate
+        TicketLastChangeTimeOlderDate
         TicketCloseTimeNewerDate
         TicketCloseTimeOlderDate
         TicketPendingTimeNewerDate