+++ scripts/test/Ticket.t
@@ -1082,6 +1082,19 @@ $Self->True(
     'TicketSearch() (HASH:TicketCreateTimeNewerMinutes => 60)',
 );
 
+# Test TicketLastChangeTimeNewerMinutes
+%TicketIDs = $TicketObject->TicketSearch(
+    Result                           => 'HASH',
+    Limit                            => 100,
+    TicketLastChangeTimeNewerMinutes => 60,
+    UserID                           => 1,
+    Permission                       => 'rw',
+);
+$Self->True(
+    $TicketIDs{$TicketID},
+    'TicketSearch() (HASH:TicketLastChangeTimeNewerMinutes => 60)',
+);
+
 # Test ArticleCreateTimeNewerMinutes
 %TicketIDs = $TicketObject->TicketSearch(
     Result                        => 'HASH',
@@ -1106,6 +1119,19 @@ $Self->True(
 $Self->False(
     $TicketIDs{$TicketID},
     'TicketSearch() (HASH:TicketCreateTimeOlderMinutes => 60)',
+);
+
+# Test TicketLastChangeOlderMinutes
+%TicketIDs = $TicketObject->TicketSearch(
+    Result                           => 'HASH',
+    Limit                            => 100,
+    TicketLastChangeTimeOlderMinutes => 60,
+    UserID                           => 1,
+    Permission                       => 'rw',
+);
+$Self->False(
+    $TicketIDs{$TicketID},
+    'TicketSearch() (HASH:TicketLastChangeTimeOlderMinutes => 60)',
 );
 
 # Test ArticleCreateOlderMinutes
@@ -1137,6 +1163,22 @@ $Self->True(
     'TicketSearch() (HASH:TicketCreateTimeNewerDate => 60)',
 );
 
+# Test TicketLastChangeTimeNewerDate
+$SystemTime = $TimeObject->SystemTime();
+%TicketIDs  = $TicketObject->TicketSearch(
+    Result                        => 'HASH',
+    Limit                         => 100,
+    TicketLastChangeTimeNewerDate => $TimeObject->SystemTime2TimeStamp(
+        SystemTime => $SystemTime - ( 60 * 60 ),
+    ),
+    UserID     => 1,
+    Permission => 'rw',
+);
+$Self->True(
+    $TicketIDs{$TicketID},
+    'TicketSearch() (HASH:TicketLastChangeTimeNewerDate => 60)',
+);
+
 # Test ArticleCreateTimeNewerDate
 $SystemTime = $TimeObject->SystemTime();
 %TicketIDs  = $TicketObject->TicketSearch(
@@ -1151,6 +1193,21 @@ $SystemTime = $TimeObject->SystemTime();
 $Self->True(
     $TicketIDs{$TicketID},
     'TicketSearch() (HASH:ArticleCreateTimeNewerDate => 60)',
+);
+
+# Test TicketLastChangeOlderDate
+%TicketIDs = $TicketObject->TicketSearch(
+    Result                        => 'HASH',
+    Limit                         => 100,
+    TicketLastChangeTimeOlderDate => $TimeObject->SystemTime2TimeStamp(
+        SystemTime => $SystemTime - ( 60 * 60 ),
+    ),
+    UserID     => 1,
+    Permission => 'rw',
+);
+$Self->False(
+    $TicketIDs{$TicketID},
+    'TicketSearch() (HASH:TicketLastChangeTimeOlderDate => 60)',
 );
 
 # Test TicketCreateOlderDate